+++ core/node/proof_data_handler/Cargo.toml
@@ -18,11 +18,10 @@ zksync_prover_interface.workspace = true
 zksync_tee_verifier.workspace = true
 zksync_types.workspace = true
 zksync_utils.workspace = true
-
-tracing.workspace = true
 anyhow.workspace = true
 axum.workspace = true
 tokio.workspace = true
+tracing.workspace = true
 
 [dev-dependencies]
 http-body-util = "0.1.0"
