+++ core/node/proof_data_handler/src/tee_request_processor.rs
@@ -7,7 +7,7 @@ use zksync_dal::{ConnectionPool, Core, CoreDal};
 use zksync_object_store::ObjectStore;
 use zksync_prover_interface::api::{
     GenericProofGenerationDataResponse, SubmitProofResponse, SubmitTeeProofRequest,
-    TeeProofGenerationDataRequest,
+    TeeProofGenerationDataRequest, TeeType,
 };
 use zksync_tee_verifier::TeeVerifierInput;
 use zksync_types::L1BatchNumber;
@@ -63,10 +63,12 @@ impl TeeRequestProcessor {
 
     pub(crate) async fn submit_proof(
         &self,
+        Path(tee_type): Path<TeeType>,
         Path(l1_batch_number): Path<u32>,
         Json(payload): Json<SubmitTeeProofRequest>,
     ) -> Result<Json<SubmitProofResponse>, RequestProcessorError> {
         tracing::info!("Received proof for block number: {:?}", l1_batch_number);
+        assert_eq!(tee_type, TeeType::Sgx, "Expected TEE type to be SGX");
 
         let l1_batch_number = L1BatchNumber(l1_batch_number);
         let mut connection = self.pool.connection().await.unwrap();
